+++ packages/subscription-controller/jest.config.cjs
@@ -14,6 +14,8 @@ module.exports = merge(baseConfig, {
   // The display name when running multiple projects
   displayName,
 
+  testEnvironment: '<rootDir>/jest.environment.cjs',
+
   // An object that configures minimum threshold enforcement for coverage results
   coverageThreshold: {
     global: {